Refactoring RecipeList with loading
