+++ src/components/RecipeList/RecipeListItem/TotalTime/TotalTime.js
@@ -1,6 +1,6 @@
 import { Chip, Tooltip } from "@mui/material";
 import AccessAlarmIcon from "@mui/icons-material/AccessAlarm";
-import { minutesToHours } from "../../../utils";
+import { minutesToHours } from "../../../../utils";
 import PropTypes from "prop-types";
 
 export const TotalTime = (props) => {
