+++ src/components/RecipeList/RecipeListItem/index.js
@@ -1,0 +1,2 @@
+import { RecipeListItem } from "./RecipeListItem";
+export default RecipeListItem;
