+++ src/components/RecipeList/index.js
@@ -1,0 +1,2 @@
+import { RecipeList } from "./RecipeList";
+export default RecipeList;
