--- src/components/RecipeListCard/index.js
@@ -1,2 +1,0 @@
-import { RecipeListCard } from "./RecipeListCard";
-export default RecipeListCard;
