+++ src/features/EditRecipe/index.js
@@ -1,0 +1,2 @@
+import { EditRecipe } from "./EditRecipe";
+export default EditRecipe;
